backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.17
websockets==13.1

# Google Cloud - pinned to compatible versions
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.20.2
google-generativeai==0.8.3
google-cloud-aiplatform==1.71.0

# Causal inference - Python 3.12 compatible
dowhy>=0.12
econml>=0.15.0
causalml>=0.15.0

# Data science
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.5.2
statsmodels==0.14.4

# DAG and graphs
networkx==3.4.2
pgmpy==0.1.26

# Jupyter notebook generation
nbformat==5.10.4
nbconvert==7.16.4

# Kaggle
kaggle==1.6.17

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.27.2
tenacity==9.0.0
structlog==24.4.0
python-json-logger==2.0.7
